miner dir
