+++ miner/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "ore-miner-web"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+gloo-worker = "0.5.0"
+log = "0.4.22"
+wasm-logger = "0.2.0"
